+++ crates/auths-infra-http/tests/cases/witness.rs
@@ -26,7 +26,7 @@ async fn start_test_server() -> (SocketAddr, WitnessServerState) {
     (addr, state)
 }
 
-fn make_test_event(prefix: &str, seq: u64) -> (Vec<u8>, Said) {
+fn make_test_event(prefix: &str, seq: u128) -> (Vec<u8>, Said) {
     let rng = ring::rand::SystemRandom::new();
     let pkcs8 = Ed25519KeyPair::generate_pkcs8(&rng).unwrap();
     let kp = Ed25519KeyPair::from_pkcs8(pkcs8.as_ref()).unwrap();
